+++ src/main/java/com/dh/ProyectoFinal/Repository/DomicilioRepository.java
@@ -3,5 +3,5 @@ package com.dh.ProyectoFinal.Repository;
 import com.dh.ProyectoFinal.Entity.Domicilio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DomicilioRepository extends JpaRepository<Domicilio,Integer> {
+public interface DomicilioRepository extends JpaRepository<Domicilio,Long> {
 }
