+++ src/main/java/com/dh/ProyectoFinal/Repository/OdontologoRepository.java
@@ -3,5 +3,5 @@ package com.dh.ProyectoFinal.Repository;
 import com.dh.ProyectoFinal.Entity.Odontologo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OdontologoRepository extends JpaRepository<Odontologo,Integer> {
+public interface OdontologoRepository extends JpaRepository<Odontologo,Long> {
 }
