+++ src/main/java/com/dh/ProyectoFinal/Repository/PacienteRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface PacienteRepository extends JpaRepository<Paciente,Integer> {
+public interface PacienteRepository extends JpaRepository<Paciente,Long> {
     Optional<Paciente> findByEmail(String email);
 }
