+++ src/main/java/com/dh/ProyectoFinal/Repository/TurnoRepository.java
@@ -3,5 +3,5 @@ package com.dh.ProyectoFinal.Repository;
 import com.dh.ProyectoFinal.Entity.Turno;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TurnoRepository extends JpaRepository<Turno,Integer> {
+public interface TurnoRepository extends JpaRepository<Turno,Long> {
 }
